src/printer_service/printer_service.py
"""
DienstplanSplitter Virtual Printer Service

This module implements a virtual printer service that:
1. Installs as a Windows printer
2. Captures print jobs and converts them to PDF
3. Uses the DienstplanSplitter core logic to split the PDFs
4. Saves the split PDFs to a configured directory
"""

import os
import sys
import win32print
import win32api
from pathlib import Path
from typing import Optional

class DienstplanPrinterService:
    """Main printer service class that handles printer installation and print jobs."""
    
    def __init__(self, printer_name: str = "Dienstplan Splitter", output_dir: Optional[Path] = None):
        """Initialize the printer service.
        
        Args:
            printer_name: Name of the virtual printer as it appears in Windows
            output_dir: Directory where split PDFs will be saved
        """
        self.printer_name = printer_name
        self.output_dir = output_dir or Path.home() / "Documents" / "Dienstplan Splitter"
        self.output_dir.mkdir(parents=True, exist_ok=True)
        
    def install_printer(self) -> bool:
        """Install the virtual printer in Windows.
        
        Returns:
            bool: True if installation was successful
        """
        try:
            # TODO: Implement printer installation using PDF Creator SDK
            return True
        except Exception as e:
            print(f"Error installing printer: {e}", file=sys.stderr)
            return False
            
    def uninstall_printer(self) -> bool:
        """Remove the virtual printer from Windows.
        
        Returns:
            bool: True if uninstallation was successful
        """
        try:
            # TODO: Implement printer removal
            return True
        except Exception as e:
            print(f"Error uninstalling printer: {e}", file=sys.stderr)
            return False
            
    def handle_print_job(self, job_id: int, document_name: str) -> bool:
        """Handle an incoming print job.
        
        Args:
            job_id: Windows print job ID
            document_name: Name of the document being printed
            
        Returns:
            bool: True if job was processed successfully
        """
        try:
            # TODO: Implement print job handling and PDF conversion
            return True
        except Exception as e:
            print(f"Error processing print job: {e}", file=sys.stderr)
            return False
            
def main():
    """Main entry point when running as a service."""
    service = DienstplanPrinterService()
    if service.install_printer():
        print(f"Successfully installed printer: {service.printer_name}")
    else:
        print("Failed to install printer", file=sys.stderr)
        sys.exit(1)

if __name__ == "__main__":
    main() 